Fix rewiew by Oleg Elifantiev (28-26)
