+++ packages/client/src/components/modal/modal.tsx
@@ -1,28 +1,24 @@
 import React, { ReactNode } from 'react'
 import './modal.pcss'
 
-type ModalProps = {
+export type ModalProps = {
   isOpen: boolean
   toggle: () => void
   children?: ReactNode
 }
 
 function Modal(props: ModalProps) {
-  return (
-    <>
-      {props.isOpen && (
-        <div
-          className="modal flex flex-ai-center flex-jc-center"
-          onClick={props.toggle}>
-          <div
-            className="modal__container"
-            onClick={event => event.stopPropagation()}>
-            {props.children}
-          </div>
-        </div>
-      )}
-    </>
-  )
+  return props.isOpen ? (
+    <div
+      className="modal flex flex-ai-center flex-jc-center"
+      onClick={props.toggle}>
+      <div
+        className="modal__container"
+        onClick={event => event.stopPropagation()}>
+        {props.children}
+      </div>
+    </div>
+  ) : null
 }
 
 export default Modal
