+++ packages/client/src/hooks/use-modal.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react'
 
 export default function useModal() {
-  const [isOpen, setisOpen] = useState(false)
+  const [isOpen, setIsOpen] = useState(false)
 
   const toggle = () => {
-    setisOpen(!isOpen)
+    setIsOpen(!isOpen)
   }
 
   return {
