+++ packages/client/src/pages/signup/signup-page.tsx
@@ -64,7 +64,6 @@ function Signup() {
     },
   ]
   const handleSubmit = async (data: unknown) => {
-
     // registration here
   }
